requirements.txt
# Core Framework
streamlit==1.50.0
pandas==2.3.3
numpy==1.26.4

# NLP & Transformers
transformers==4.56.2
torch==2.8.0
scikit-learn==1.7.2
nltk==3.9.1

# Medical NLP
spacy==3.7.5
scispacy==0.5.5
en_core_sci_sm @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_sm-0.5.4.tar.gz

# Optional medspaCy tools (you can comment these if not using medspacy)
medspacy==1.3.1
medspacy_quickumls==3.2
medspacy_unqlite==0.9.8

# QA, NER & zero-shot
sentencepiece==0.2.0
accelerate==1.10.1
huggingface-hub==0.35.3
safetensors==0.6.2

# Google Gemini Integration
google-generativeai==0.8.5
google-api-python-client==2.183.0
protobuf==5.29.5
google-auth==2.41.1

# Utilities
regex==2025.9.18
tqdm==4.67.1
requests==2.32.5
pydantic==2.11.9
python-dateutil==2.9.0.post0
PyYAML==6.0.3

# Logging / Display
rich==14.1.0
loguru==0.7.3

# Optional visualization
altair==5.5.0
pydeck==0.9.1
